+++ tests/fscript_test.cc
@@ -3229,3 +3229,23 @@ TEST(FScript, consts) {
 
   TK_OBJECT_UNREF(obj);
 }
+
+TEST(FScript, levelize) {
+  value_t v;
+  value_t s;
+  tk_object_t* obj = object_default_create();
+  
+  fscript_eval(obj, "var t = levelize('0-20;21-40;41-60;61-80;81-100', 10); t", &v);
+  ASSERT_EQ(value_int(&v), 0);
+
+  fscript_eval(obj, "var t = levelize('0-20;21-40;41-60;61-80;81-100', 30); t", &v);
+  ASSERT_EQ(value_int(&v), 1);
+  
+  fscript_eval(obj, "var t = levelize('0-20;21-40;41-60;61-80;81-100', 200); t", &v);
+  ASSERT_EQ(value_int(&v), 4);
+
+  value_reset(&v);
+  
+  TK_OBJECT_UNREF(obj);
+}
+